put auth logic in middleware fn
