+++ controllers/completionsController.js
@@ -2,7 +2,6 @@ require('dotenv').config();
 const path = require('path');
 const Prefixy = require(path.resolve(path.dirname(__dirname), 'prefixy'));
 const _ = require('lodash');
-const jwt = require("jsonwebtoken");
 
 const formatCompletionsWithScores = completions => {
   return _.chunk(completions, 2).map(completion => (
@@ -13,19 +12,8 @@ const formatCompletionsWithScores = completions => {
   ));
 };
 
-const resolveTenant = token => {
-  Prefixy.tenant = jwt.verify(token, process.env.SECRET).tenant;
-};
-
 module.exports = {
   get: async function(req, res, next) {
-    try {
-      resolveTenant(req.query.token);
-    } catch(error) {
-      error.status = 401;
-      return next(error);
-    }
-
     const prefix = req.query.prefix;
     const opts = {
       limit: req.query.limit || Prefixy.suggestionCount,
@@ -47,13 +35,6 @@ module.exports = {
   },
 
   post: function(req, res, next) {
-    try {
-      resolveTenant(req.body.token);
-    } catch(error) {
-      error.status = 401;
-      return next(error);
-    }
-
     const completions = req.body.completions;
 
     Prefixy.invoke(() => Prefixy.insertCompletions(completions));
@@ -62,13 +43,6 @@ module.exports = {
   },
 
   delete: function(req, res, next) {
-    try {
-      resolveTenant(req.body.token);
-    } catch(error) {
-      error.status = 401;
-      return next(error);
-    }
-
     const completions = req.body.completions;
 
     Prefixy.invoke(() => Prefixy.deleteCompletions(completions));
